require 'sinatra/base' -> this is key.
This is huge for extensions to do.  Requiring just `sinatra`
will cause all public methods to be installed as a delegate
on (main) which is bad for those not using Sinatra classic
style (see http://www.sinatrarb.com/extensions.html).

This really put me through a loop while testing; it
was not using the `mongo` setting but the default database
which caused my tests to fail without much help to why.

Please pull.  I wouldn't want this to happen to others.

Great work on this!  Thank you.